MIMIC Tables

Collection of functions to create prearranged views of MIMIC patient data for ReviewR.

mimic3_table_all_patients(table_map, db_connection)

mimic3_table_admissions(table_map, db_connection, subject_id)

mimic3_table_callout(table_map, db_connection, subject_id)

mimic3_table_chart_events(table_map, db_connection, subject_id)

mimic3_table_cpt_events(table_map, db_connection, subject_id)

mimic3_table_diagnoses_icd(table_map, db_connection, subject_id)

mimic3_table_drg_codes(table_map, db_connection, subject_id)

mimic3_table_icu_stays(table_map, db_connection, subject_id)

mimic3_table_lab_events(table_map, db_connection, subject_id)

mimic3_table_microbiology_events(table_map, db_connection, subject_id)

mimic3_table_note_events(table_map, db_connection, subject_id)

mimic3_table_prescriptions(table_map, db_connection, subject_id)

mimic3_table_procedure_events(table_map, db_connection, subject_id)

mimic3_table_procedures_icd(table_map, db_connection, subject_id)

mimic3_table_services(table_map, db_connection, subject_id)

mimic3_table_transfers(table_map, db_connection, subject_id)

Arguments

table_map
tibble containing a the CDM that most closely matches the user's database and a map of standard tables to user tables

db_connection
Connection info received from the database setup module

subject_id
The selected subject