Demo SQLite Database Module

This module will create an in memory SQLite database with demo data from the CMS 2008-2010 Data Entrepreneurs’ Synthetic Public Use File (DE-SynPUF) from OHDSI. It will allow you to preview the functionality of ReviewR if you do not have access to a database of patient information.

This module consists of the following components:

Module UI function

demo_sqlite_setup_ui: A uiOutput that allows users to connect to the demonstration database.

Module Server function

demo_sqlite_setup_server: The logic that creates the demonstration SQLite database and returns a DBI::dbConnect() object used to connect to the demo database.

Usage

demo_sqlite_setup_ui(id)

demo_sqlite_setup_server(id)

Arguments

id

The module namespace

Value

demo_sqlite_setup_ui:

tagList

The Demo SQLite Setup UI

demo_sqlite_setup_server:

reactiveValues

moduleName: A string, containing the module moniker.

moduleType: A string, with the module type (what does it do?)

setup_ui: The module setup ui function

is_connected: A string, with module connection status. Valid statuses are 'yes' or 'no'.

db_con: A DBI::dbConnect object, containing the user configured connection information.