Chart Review Interface Module

This module determines if "View" or "Review" interface is needed based on presence or absence of a configured abstraction module. When an abstraction module is configured on the "Setup" tab of ReviewR, a column will be created to the right of the patient chart for the abstraction data collection instrument. Otherwise, the patient chart will take up the full view.

This module consists of the following components:

Module UI function

chartreview_ui: A uiOutput containing the View or Review interface

Module Server function

chartreview_server: The logic that controls the layout of the "Chart Review" tab in ReviewR. "View" if no abstraction module is configured. "Review" if abstraction module configured.

Usage

chartreview_ui(id)

chartreview_server(id, database_vars, abstract_vars)

Arguments

id

The module namespace

database_vars

A reactiveValues object as returned by mod_database_setup.

abstract_vars

A reactiveValues object as returned by mod_abstraction_setup.

Value

chartreview_ui:

tagList

The Chart Review UI

chartreview_server:

NULL

This function has no return, other than creating a UI output for the UI function of this module.