REDCap Module

This module contains all of the REDCap setup and instrument upload/download components.

redcap_connect_ui(id)

redcap_connect_logic(input, output, session)

redcap_initialize_logic(input, output, session, rc_url, rc_token)

redcap_connected_ui(id)

redcap_connected_logic(input, output, session, connect, rc_project_info)

redcap_instrument_config_ui(id)

redcap_instrument_select_logic(
  input,
  output,
  session,
  rc_connect_press,
  rc_connection
)

redcap_instrument_config_logic(
  input,
  output,
  session,
  rc_connection,
  instruments,
  instrument_selection
)

redcap_instrument_config_reviewer_logic(
  input,
  output,
  session,
  rc_instrument,
  rc_identifier,
  rc_connection
)

rc_instrument_configured_ui(id)

rc_instrument_configured_logic(
  input,
  output,
  session,
  rc_instrument_info,
  selected_instrument
)

redcap_instrument_ui(id)

redcap_instrument_logic(
  input,
  output,
  session,
  rc_connection,
  instruments,
  instrument_selection,
  rc_instrument,
  rc_identifier,
  rc_reviewer,
  rc_selected_reviewer,
  subject_id,
  reviewr_upload_btn,
  reviewr_connect_btn
)

upload_redcap_logic(
  input,
  output,
  session,
  rc_connection,
  rc_recordID,
  rc_instrument,
  instrumentData,
  previousData,
  currentSubject,
  rc_upload_btn_press,
  abstraction_complete,
  abstraction_complete_val,
  all_instruments,
  instrument_selection
)

Arguments

id
The namespace id for the UI output

input
internal

output
internal

session
internal

rc_url
A character string containing the REDCap API URL for your institution ex: https://redcap.ucdenver.edu/api/

rc_token
A password string representing your REDCap API token

connect
A press of the REDCap connect button from the Setup Tab

rc_project_info
The project information associated with your REDCap API key

rc_connect_press
A press of the REDCap connect button from the Setup Tab

rc_connection
A REDCap API connection object

instruments
All REDCap instruments associated with the project API key

instrument_selection
The user selected REDCap Instrument

rc_instrument
The user selected REDCap Instrument

rc_identifier
The field that contains the REDCap Identifier

rc_instrument_info
Information about the configured REDCap Instrument

selected_instrument
The currently selected REDCap Instrument

rc_reviewer
The REDCap field that holds reviewer names

rc_selected_reviewer
The selected reviewer from the list of previous reviewers (if present)

subject_id
The currently selected patient identifier

reviewr_upload_btn
A button press of the REDCap upload button

reviewr_connect_btn
A button press of the REDCap Connect button

rc_recordID
The REDCap Record ID for the currently selected subject

instrumentData
The user entered data currently present in the REDCap instrument

previousData
Previous Abstraction "Instrument Complete" Value

currentSubject
Abstraction data associated with the currently selected subject

rc_upload_btn_press
A press of the Upload to REDCap button

abstraction_complete
Person readable abstraction complete value

abstraction_complete_val
Machine readable abstraction complete value

all_instruments
All data abstraction instruments in the project