Render REDCap Instrument

Collection of functions to map REDCap question types to native Shiny widgets.

reviewr_textInput(id, field_label, value = NULL, placeholder = NULL, ...)

reviewr_dateInput(id, field_label, value = NULL, ...)

reviewr_dropdown(id, field_label, required, choices, value = NULL, ...)

reviewr_truefalse(id, field_label, required, value = NULL, ...)

reviewr_yesno(id, field_label, required, value = NULL, ...)

reviewr_radio(id, field_label, required, choices, value = NULL, ...)

reviewr_checkbox(id, field_label, choices, value = NULL, ...)

reviewr_notes(id, field_label, value = NULL, ...)

reviewr_integer(id, field_label, value = NULL, ...)

render_redcap(
  reviewr_type,
  field_name,
  field_label,
  required,
  choices,
  current_subject_data = NULL
)

Arguments

id
Unique REDCap question identifier

field_label
Question text, with formatting

value
Default value or previous data if question has previously been answered

placeholder
Placeholder text to help a reviewer decide how to answer the question

...
Any additional parameters

required
Is this a required REDCap question type?

choices
REDCap choices for the question.

current_subject_data
Previous REDCap data on the current subject