ReviewR Supported Database Schemas

A dataset containing data model information along with the corresponding version and schema.

supported_models

Format

A data frame with 12 rows and 4 variables:

file_path

Where schema was imported from

data_model

Data model name, OMOP or MIMIC

model_version

Version of the data model

data

Nested database schemas, including included table and field mappings

...
Source

https://github.com/OHDSI/CommonDataModel/

https://github.com/MIT-LCP/mimic-code