care_site

Table Description

The CARE_SITE table contains a list of uniquely identified institutional (physical or organizational) units where healthcare delivery is practiced (offices, wards, hospitals, clinics, etc.).

ETL Conventions

Care site is a unique combination of location_id and place_of_service_source_value. Care site does not take into account the provider (human) information such a specialty. Many source data do not make a distinction between individual and institutional providers. The CARE_SITE table contains the institutional providers. If the source, instead of uniquely identifying individual Care Sites, only provides limited information such as Place of Service, generic or “pooled” Care Site records are listed in the CARE_SITE table. There can be hierarchical and business relationships between Care Sites. For example, wards can belong to clinics or departments, which can in turn belong to hospitals, which in turn can belong to hospital systems, which in turn can belong to HMOs.The relationships between Care Sites are defined in the FACT_RELATIONSHIP table.

Table Description

The Standardized Vocabularies contains records, or Concepts, that uniquely identify each fundamental unit of meaning used to express clinical information in all domain tables of the CDM. Concepts are derived from vocabularies, which represent clinical information across a domain (e.g. conditions, drugs, procedures) through the use of codes and associated descriptions. Some Concepts are designated Standard Concepts, meaning these Concepts can be used as normative expressions of a clinical entity within the OMOP Common Data Model and within standardized analytics. Each Standard Concept belongs to one domain, which defines the location where the Concept would be expected to occur within data tables of the CDM.

Concepts can represent broad categories (like ‘Cardiovascular disease’), detailed clinical elements (‘Myocardial infarction of the anterolateral wall’) or modifying characteristics and attributes that define Concepts at various levels of detail (severity of a disease, associated morphology, etc.).

Records in the Standardized Vocabularies tables are derived from national or international vocabularies such as SNOMED-CT, RxNorm, and LOINC, or custom Concepts defined to cover various aspects of observational data analysis.

Table Description

The CONCEPT_CLASS table is a reference table, which includes a list of the classifications used to differentiate Concepts within a given Vocabulary. This reference table is populated with a single record for each Concept Class.

Table Description

The CONCEPT_SYNONYM table is used to store alternate names and descriptions for Concepts.

Table Description

A Condition Era is defined as a span of time when the Person is assumed to have a given condition. Similar to Drug Eras, Condition Eras are chronological periods of Condition Occurrence. Combining individual Condition Occurrences into a single Condition Era serves two purposes:

It allows aggregation of chronic conditions that require frequent ongoing care, instead of treating each Condition Occurrence as an independent event.

It allows aggregation of multiple, closely timed doctor visits for the same Condition to avoid double-counting the Condition Occurrences. For example, consider a Person who visits her Primary Care Physician (PCP) and who is referred to a specialist. At a later time, the Person visits the specialist, who confirms the PCP’s original diagnosis and provides the appropriate treatment to resolve the condition. These two independent doctor visits should be aggregated into one Condition Era.

ETL Conventions

Each Condition Era corresponds to one or many Condition Occurrence records that form a continuous interval. The condition_concept_id field contains Concepts that are identical to those of the CONDITION_OCCURRENCE table records that make up the Condition Era. In contrast to Drug Eras, Condition Eras are not aggregated to contain Conditions of different hierarchical layers. The SQl Script for generating CONDITION_ERA records can be found here The Condition Era Start Date is the start date of the first Condition Occurrence. The Condition Era End Date is the end date of the last Condition Occurrence. Condition Eras are built with a Persistence Window of 30 days, meaning, if no occurrence of the same condition_concept_id happens within 30 days of any one occurrence, it will be considered the condition_era_end_date.

Table Description

This table contains records of Events of a Person suggesting the presence of a disease or medical condition stated as a diagnosis, a sign, or a symptom, which is either observed by a Provider or reported by the patient.

User Guide

Conditions are defined by Concepts from the Condition domain, which form a complex hierarchy. As a result, the same Person with the same disease may have multiple Condition records, which belong to the same hierarchical family. Most Condition records are mapped from diagnostic codes, but recorded signs, symptoms and summary descriptions also contribute to this table. Rule out diagnoses should not be recorded in this table, but in reality their negating nature is not always captured in the source data, and other precautions must be taken when when identifying Persons who should suffer from the recorded Condition. Record all conditions as they exist in the source data. Any decisions about diagnosis/phenotype definitions would be done through cohort specifications. These cohorts can be housed in the COHORT table. Conditions span a time interval from start to end, but are typically recorded as single snapshot records with no end date. The reason is twofold: (i) At the time of the recording the duration is not known and later not recorded, and (ii) the Persons typically cease interacting with the healthcare system when they feel better, which leads to incomplete capture of resolved Conditions. The CONDITION_ERA table addresses this issue. Family history and past diagnoses (‘history of’) are not recorded in this table. Instead, they are listed in the OBSERVATION table. Codes written in the process of establishing the diagnosis, such as ‘question of’ of and ‘rule out’, should not represented here. Instead, they should be recorded in the OBSERVATION table, if they are used for analyses. However, this information is not always available.

Table Description

The death domain contains the clinical event for how and when a Person dies. A person can have up to one record if the source system contains evidence about the Death, such as: Condition in an administrative claim, status of enrollment into a health plan, or explicit record in EHR data.

Table Description

The Device domain captures information about a person’s exposure to a foreign physical object or instrument which is used for diagnostic or therapeutic purposes through a mechanism beyond chemical action. Devices include implantable objects (e.g. pacemakers, stents, artificial joints), medical equipment and supplies (e.g. bandages, crutches, syringes), other instruments used in medical procedures (e.g. sutures, defibrillators) and material used in clinical care (e.g. adhesives, body material, dental material, surgical material).

User Guide

The distinction between Devices or supplies and Procedures are sometimes blurry, but the former are physical objects while the latter are actions, often to apply a Device or supply.

ETL Conventions

Source codes and source text fields mapped to Standard Concepts of the Device Domain have to be recorded here.

Table Description

The DOMAIN table includes a list of OMOP-defined Domains the Concepts of the Standardized Vocabularies can belong to. A Domain defines the set of allowable Concepts for the standardized fields in the CDM tables. For example, the “Condition” Domain contains Concepts that describe a condition of a patient, and these Concepts can only be stored in the condition_concept_id field of the CONDITION_OCCURRENCE and CONDITION_ERA tables. This reference table is populated with a single record for each Domain and includes a descriptive name for the Domain.

Table Description

A Drug Era is defined as a span of time when the Person is assumed to be exposed to a particular active ingredient. A Drug Era is not the same as a Drug Exposure: Exposures are individual records corresponding to the source when Drug was delivered to the Person, while successive periods of Drug Exposures are combined under certain rules to produce continuous Drug Eras.

ETL Conventions

The SQL script for generating DRUG_ERA records can be found here.

Table Description

The Device domain captures information about a person’s exposure to a foreign physical object or instrument which is used for diagnostic or therapeutic purposes through a mechanism beyond chemical action. Devices include implantable objects (e.g. pacemakers, stents, artificial joints), medical equipment and supplies (e.g. bandages, crutches, syringes), other instruments used in medical procedures (e.g. sutures, defibrillators) and material used in clinical care (e.g. adhesives, body material, dental material, surgical material).

User Guide

The distinction between Devices or supplies and Procedures are sometimes blurry, but the former are physical objects while the latter are actions, often to apply a Device or supply.

ETL Conventions

Source codes and source text fields mapped to Standard Concepts of the Device Domain have to be recorded here.

Table Description

The DRUG_STRENGTH table contains structured content about the amount or concentration and associated units of a specific ingredient contained within a particular drug product. This table is supplemental information to support standardized analysis of drug utilization.

Table Description

The MEASUREMENT table contains records of Measurements, i.e. structured values (numerical or categorical) obtained through systematic and standardized examination or testing of a Person or Person’s sample. The MEASUREMENT table contains both orders and results of such Measurements as laboratory tests, vital signs, quantitative findings from pathology reports, etc. Measurements are stored as attribute value pairs, with the attribute as the Measurement Concept and the value representing the result. The value can be a Concept (stored in VALUE_AS_CONCEPT), or a numerical value (VALUE_AS_NUMBER) with a Unit (UNIT_CONCEPT_ID). The Procedure for obtaining the sample is housed in the PROCEDURE_OCCURRENCE table, though it is unnecessary to create a PROCEDURE_OCCURRENCE record for each measurement if one does not exist in the source data. Measurements differ from Observations in that they require a standardized test or some other activity to generate a quantitative or qualitative result. If there is no result, it is assumed that the lab test was conducted but the result was not captured.

User Guide

Measurements are predominately lab tests with a few exceptions, like blood pressure or function tests. Results are given in the form of a value and unit combination. When investigating measurements, look for operator_concept_ids (<, >, etc.).

ETL Conventions

Only records where the source value maps to a Concept in the measurement domain should be included in this table. Even though each Measurement always has a result, the fields VALUE_AS_NUMBER and VALUE_AS_CONCEPT_ID are not mandatory as often the result is not given in the source data. When the result is not known, the Measurement record represents just the fact that the corresponding Measurement was carried out, which in itself is already useful information for some use cases. For some Measurement Concepts, the result is included in the test. For example, ICD10 CONCEPT_ID 45548980 ‘Abnormal level of unspecified serum enzyme’ indicates a Measurement and the result (abnormal). In those situations, the CONCEPT_RELATIONSHIP table in addition to the ‘Maps to’ record contains a second record with the relationship_id set to ‘Maps to value’. In this example, the ‘Maps to’ relationship directs to 4046263 ‘Enzyme measurement’ as well as a ‘Maps to value’ record to 4135493 ‘Abnormal’.

Table Description

The NOTE table captures unstructured information that was recorded by a provider about a patient in free text (in ASCII, or preferably in UTF8 format) notes on a given date. The type of note_text is CLOB or varchar(MAX) depending on RDBMS.

ETL Conventions

HL7/LOINC CDO is a standard for consistent naming of documents to support a range of use cases: retrieval, organization, display, and exchange. It guides the creation of LOINC codes for clinical notes. CDO annotates each document with 5 dimensions:

Kind of Document: Characterizes the general structure of the document at a macro level (e.g. Anesthesia Consent)

Type of Service: Characterizes the kind of service or activity (e.g. evaluations, consultations, and summaries). The notion of time sequence, e.g., at the beginning (admission) at the end (discharge) is subsumed in this axis. Example: Discharge Teaching.

Setting: Setting is an extension of CMS’s definitions (e.g. Inpatient, Outpatient)

Subject Matter Domain (SMD): Characterizes the subject matter domain of a note (e.g. Anesthesiology)

Role: Characterizes the training or professional level of the author of the document, but does not break down to specialty or subspecialty (e.g. Physician) Each combination of these 5 dimensions rolls up to a unique LOINC code.

According to CDO requirements, only 2 of the 5 dimensions are required to properly annotate a document; Kind of Document and any one of the other 4 dimensions. However, not all the permutations of the CDO dimensions will necessarily yield an existing LOINC code. Each of these dimensions are contained in the OMOP Vocabulary under the domain of ‘Meas Value’ with each dimension represented as a Concept Class.

Table Description

The OBSERVATION table captures clinical facts about a Person obtained in the context of examination, questioning or a procedure. Any data that cannot be represented by any other domains, such as social and lifestyle facts, medical history, family history, etc. are recorded here.

User Guide

Observations differ from Measurements in that they do not require a standardized test or some other activity to generate clinical fact. Typical observations are medical history, family history, the stated need for certain treatment, social circumstances, lifestyle choices, healthcare utilization patterns, etc. If the generation clinical facts requires a standardized testing such as lab testing or imaging and leads to a standardized result, the data item is recorded in the MEASUREMENT table. If the clinical fact observed determines a sign, symptom, diagnosis of a disease or other medical condition, it is recorded in the CONDITION_OCCURRENCE table. Valid Observation Concepts are not enforced to be from any domain though they still should be Standard Concepts.

ETL Conventions

Records whose Source Values map to any domain besides Condition, Procedure, Drug, Measurement or Device should be stored in the Observation table. Observations can be stored as attribute value pairs, with the attribute as the Observation Concept and the value representing the clinical fact. This fact can be a Concept (stored in VALUE_AS_CONCEPT), a numerical value (VALUE_AS_NUMBER), a verbatim string (VALUE_AS_STRING), or a datetime (VALUE_AS_DATETIME). Even though Observations do not have an explicit result, the clinical fact can be stated separately from the type of Observation in the VALUE_AS_* fields. It is recommended for Observations that are suggestive statements of positive assertion should have a value of ‘Yes’ (concept_id=4188539), recorded, even though the null value is the equivalent.

Table Description

This table contains records which define spans of time during which two conditions are expected to hold: (i) Clinical Events that happened to the Person are recorded in the Event tables, and (ii) absense of records indicate such Events did not occur during this span of time.

User Guide

For each Person, one or more OBSERVATION_PERIOD records may be present, but they will not overlap or be back to back to each other. Events may exist outside all of the time spans of the OBSERVATION_PERIOD records for a patient, however, absence of an Event outside these time spans cannot be construed as evidence of absence of an Event. Incidence or prevalence rates should only be calculated for the time of active OBSERVATION_PERIOD records. When constructing cohorts, outside Events can be used for inclusion criteria definition, but without any guarantee for the performance of these criteria. Also, OBSERVATION_PERIOD records can be as short as a single day, greatly disturbing the denominator of any rate calculation as part of cohort characterizations. To avoid that, apply minimal observation time as a requirement for any cohort definition.

ETL Conventions

Each Person needs to have at least one OBSERVATION_PERIOD record, which should represent time intervals with a high capture rate of Clinical Events. Some source data have very similar concepts, such as enrollment periods in insurance claims data. In other source data such as most EHR systems these time spans need to be inferred under a set of assumptions. It is the discretion of the ETL developer to define these assumptions. In many ETL solutions the start date of the first occurrence or the first high quality occurrence of a Clinical Event (Condition, Drug, Procedure, Device, Measurement, Visit) is defined as the start of the OBSERVATION_PERIOD record, and the end date of the last occurrence of last high quality occurrence of a Clinical Event, or the end of the database period becomes the end of the OBSERVATOIN_PERIOD for each Person. If a Person only has a single Clinical Event the OBSERVATION_PERIOD record can be as short as one day. Depending on these definitions it is possible that Clinical Events fall outside the time spans defined by OBSERVATION_PERIOD records. Family history or history of Clinical Events generally are not used to generate OBSERVATION_PERIOD records around the time they are referring to. Any two overlapping or adjacent OBSERVATION_PERIOD records have to be merged into one.

Table Description

The PAYER_PLAN_PERIOD table captures details of the period of time that a Person is continuously enrolled under a specific health Plan benefit structure from a given Payer. Each Person receiving healthcare is typically covered by a health benefit plan, which pays for (fully or partially), or directly provides, the care. These benefit plans are provided by payers, such as health insurances or state or government agencies. In each plan the details of the health benefits are defined for the Person or her family, and the health benefit Plan might change over time typically with increasing utilization (reaching certain cost thresholds such as deductibles), plan availability and purchasing choices of the Person. The unique combinations of Payer organizations, health benefit Plans and time periods in which they are valid for a Person are recorded in this table.

User Guide

A Person can have multiple, overlapping, Payer_Plan_Periods in this table. For example, medical and drug coverage in the US can be represented by two Payer_Plan_Periods. The details of the benefit structure of the Plan is rarely known, the idea is just to identify that the Plans are different.

Table Description

This table serves as the central identity management for all Persons in the database. It contains records that uniquely identify each person or patient, and some demographic information.

User Guide

All records in this table are independent Persons.

ETL Conventions

All Persons in a database needs one record in this table, unless they fail data quality requirements specified in the ETL. Persons with no Events should have a record nonetheless. If more than one data source contributes Events to the database, Persons must be reconciled, if possible, across the sources to create one single record per Person. The content of the BIRTH_DATETIME must be equivalent to the content of BIRTH_DAY, BIRTH_MONTH and BIRTH_YEAR.

Table Description

This table contains records of activities or processes ordered by, or carried out by, a healthcare provider on the patient with a diagnostic or therapeutic purpose.

User Guide

Lab tests are not a procedure, if something is observed with an expected resulting amount and unit then it should be a measurement. Phlebotomy is a procedure but so trivial that it tends to be rarely captured. It can be assumed that there is a phlebotomy procedure associated with many lab tests, therefore it is unnecessary to add them as separate procedures. If the user finds the same procedure over concurrent days, it is assumed those records are part of a procedure lasting more than a day. This logic is in lieu of the procedure_end_date, which will be added in a future version of the CDM.

ETL Conventions

If a procedure lasts more than a day, then it should be recorded as a separate record for each day the procedure occurred, this logic is in lieu of the PROCEDURE_END_DATE, which will be added in a future version of the CDM. When dealing with duplicate records, the ETL must determine whether to sum them up into one record or keep them separate. Things to consider are: - Same Procedure - Same PROCEDURE_DATETIME - Same Visit Occurrence or Visit Detail - Same Provider - Same Modifier for Procedures. Source codes and source text fields mapped to Standard Concepts of the Procedure Domain have to be recorded here.

Table Description

The PROVIDER table contains a list of uniquely identified healthcare providers. These are individuals providing hands-on healthcare to patients, such as physicians, nurses, midwives, physical therapists etc.

User Guide

Many sources do not make a distinction between individual and institutional providers. The PROVIDER table contains the individual providers. If the source, instead of uniquely identifying individual providers, only provides limited information such as specialty, generic or ‘pooled’ Provider records are listed in the PROVIDER table.

Table Description

The RELATIONSHIP table provides a reference list of all types of relationships that can be used to associate any two concepts in the CONCEPT_RELATIONSHP table.

Table Description

This table contains Events where Persons engage with the healthcare system for a duration of time. They are often also called “Encounters”. Visits are defined by a configuration of circumstances under which they occur, such as (i) whether the patient comes to a healthcare institution, the other way around, or the interaction is remote, (ii) whether and what kind of trained medical staff is delivering the service during the Visit, and (iii) whether the Visit is transient or for a longer period involving a stay in bed.

User Guide

The configuration defining the Visit are described by Concepts in the Visit Domain, which form a hierarchical structure, but rolling up to generally familiar Visits adopted in most healthcare systems worldwide:

Inpatient Visit: Person visiting hospital, at a Care Site, in bed, for duration of more than one day, with physicians and other Providers permanently available to deliver service around the clock

Emergency Room Visit: Person visiting dedicated healthcare institution for treating emergencies, at a Care Site, within one day, with physicians and Providers permanently available to deliver service around the clock

Emergency Room and Inpatient Visit: Person visiting ER followed by a subsequent Inpatient Visit, where Emergency department is part of hospital, and transition from the ER to other hospital departments is undefined

Non-hospital institution Visit: Person visiting dedicated institution for reasons of poor health, at a Care Site, long-term or permanently, with no physician but possibly other Providers permanently available to deliver service around the clock

Outpatient Visit: Person visiting dedicated ambulatory healthcare institution, at a Care Site, within one day, without bed, with physicians or medical Providers delivering service during Visit

Home Visit: Provider visiting Person, without a Care Site, within one day, delivering service

Telehealth Visit: Patient engages with Provider through communication media

Pharmacy Visit: Person visiting pharmacy for dispensing of Drug, at a Care Site, within one day

Laboratory Visit: Patient visiting dedicated institution, at a Care Site, within one day, for the purpose of a Measurement.

Ambulance Visit: Person using transportation service for the purpose of initiating one of the other Visits, without a Care Site, within one day, potentially with Providers accompanying the Visit and delivering service

Case Management Visit: Person interacting with healthcare system, without a Care Site, within a day, with no Providers involved, for administrative purposes

The Visit duration, or ‘length of stay’, is defined as VISIT_END_DATE - VISIT_START_DATE. For all Visits this is <1 day, except Inpatient Visits and Non-hospital institution Visits. The CDM also contains the VISIT_DETAIL table where additional information about the Visit is stored, for example, transfers between units during an inpatient Visit.

ETL Conventions

Visits can be derived easily if the source data contain coding systems for Place of Service or Procedures, like CPT codes for well visits. In those cases, the codes can be looked up and mapped to a Standard Visit Concept. Otherwise, Visit Concepts have to be identified in the ETL process. This table will contain concepts in the Visit domain. These concepts are arranged in a hierarchical structure to facilitate cohort definitions by rolling up to generally familiar Visits adopted in most healthcare systems worldwide. Visits can be adjacent to each other, i.e. the end date of one can be identical with the start date of the other. As a consequence, more than one-day Visits or their descendants can be recorded for the same day. Multi-day visits must not overlap, i.e. share days other than start and end days. It is often the case that some logic should be written for how to define visits and how to assign Visit_Concept_Id. For example, in US claims outpatient visits that appear to occur within the time period of an inpatient visit can be rolled into one with the same Visit_Occurrence_Id. In EHR data inpatient visits that are within one day of each other may be strung together to create one visit. It will all depend on the source data and how encounter records should be translated to visit occurrences. Providers can be associated with a Visit through the PROVIDER_ID field, or indirectly through PROCEDURE_OCCURRENCE records linked both to the VISIT and PROVIDER tables.

Table Description

The VOCABULARY table includes a list of the Vocabularies collected from various sources or created de novo by the OMOP community. This reference table is populated with a single record for each Vocabulary source and includes a descriptive name and other associated attributes for the Vocabulary.

care_site

concept

concept_class

concept_synonym

condition_era

condition_occurrence

death

device_exposure

domain

drug_era

drug_exposure

drug_strength

measurement

note

observation

observation_period

payer_plan_period

person

procedure_occurrence

provider

relationship

visit_occurrence

vocabulary

Format

A data frame with 0 rows and 6 variables:

care_site_id

character

care_site_name

character

place_of_service_concept_id

character

location_id

character

care_site_source_value

character

place_of_service_source_value

character

...

A data frame with 2355 rows and 10 variables:

concept_name

character

domain_id

character

vocabulary_id

character

concept_class_id

character

standard_concept

character

concept_code

character

valid_start_date

Date

valid_end_date

Date

invalid_reason

character

concept_id

integer

...

A data frame with 0 rows and 3 variables:

concept_class_id

character

concept_class_name

character

concept_class_concept_id

character

...

A data frame with 5835 rows and 2 variables:

concept_synonym_name

character

concept_id

integer

...

A data frame with 688 rows and 6 variables:

condition_era_id

integer

person_id

integer

condition_concept_id

integer

condition_era_start_date

Date

condition_era_end_date

Date

condition_occurrence_count

integer

...

A data frame with 992 rows and 15 variables:

condition_occurrence_id

integer

person_id

integer

condition_concept_id

integer

condition_start_date

Date

condition_start_datetime

POSIXct

condition_end_date

Date

condition_end_datetime

POSIXct

condition_type_concept_id

integer

stop_reason

character

provider_id

integer

visit_occurrence_id

integer

condition_source_value

character

condition_source_concept_id

integer

condition_status_source_value

character

condition_status_concept_id

integer

...

A data frame with 2 rows and 7 variables:

person_id

integer

death_date

Date

death_datetime

POSIXct

death_type_concept_id

integer

cause_concept_id

integer

cause_source_value

character

cause_source_concept_id

integer

...

A data frame with 4 rows and 14 variables:

device_exposure_id

integer

person_id

integer

device_concept_id

integer

device_exposure_start_date

Date

device_exposure_start_datetime

POSIXct

device_exposure_end_date

Date

device_exposure_end_datetime

POSIXct

device_type_concept_id

integer

unique_device_id

character

quantity

integer

provider_id

integer

visit_occurrence_id

integer

device_source_value

character

device_source_concept_id

integer

...

A data frame with 0 rows and 3 variables:

domain_id

character

domain_name

character

domain_concept_id

character

...

A data frame with 344 rows and 7 variables:

drug_era_id

integer

person_id

integer

drug_concept_id

integer

drug_era_start_date

Date

drug_era_end_date

Date

drug_exposure_count

integer

gap_days

integer

...

A data frame with 363 rows and 22 variables:

drug_exposure_id

integer

person_id

integer

drug_concept_id

integer

drug_exposure_start_date

Date

drug_exposure_start_datetime

POSIXct

drug_exposure_end_date

Date

drug_exposure_end_datetime

POSIXct

verbatim_end_date

Date

drug_type_concept_id

integer

stop_reason

character

refills

integer

quantity

numeric

days_supply

integer

sig

character

route_concept_id

integer

lot_number

character

provider_id

integer

visit_occurrence_id

integer

drug_source_value

character

drug_source_concept_id

integer

route_source_value

character

dose_unit_source_value

character

...

A data frame with 475035 rows and 9 variables:

amount_value

numeric

numerator_value

numeric

denominator_value

numeric

box_size

integer

valid_start_date

Date

valid_end_date

Date

invalid_reason

character

ingredient_concept_id

integer

drug_concept_id

integer

...

A data frame with 194 rows and 18 variables:

measurement_id

integer

person_id

integer

measurement_concept_id

integer

measurement_date

Date

measurement_datetime

POSIXct

measurement_type_concept_id

integer

operator_concept_id

integer

value_as_number

numeric

value_as_concept_id

integer

unit_concept_id

integer

range_low

numeric

range_high

numeric

provider_id

integer

visit_occurrence_id

integer

measurement_source_value

character

measurement_source_concept_id

integer

unit_source_value

character

value_source_value

character

...

A data frame with 851 rows and 13 variables:

note_id

integer

person_id

integer

note_date

Date

note_datetime

POSIXct

note_type_concept_id

integer

note_class_concept_id

integer

note_title

character

note_text

character

encoding_concept_id

integer

language_concept_id

integer

provider_id

integer

visit_occurrence_id

integer

note_source_value

character

...

A data frame with 184 rows and 17 variables:

observation_id

integer

person_id

integer

observation_concept_id

integer

observation_date

Date

observation_datetime

POSIXct

observation_type_concept_id

integer

value_as_number

numeric

value_as_string

character

value_as_concept_id

integer

qualifier_concept_id

integer

unit_concept_id

integer

provider_id

integer

visit_occurrence_id

integer

observation_source_value

character

observation_source_concept_id

integer

unit_source_value

character

qualifier_source_value

character

...

A data frame with 8 rows and 5 variables:

observation_period_id

integer

person_id

integer

observation_period_start_date

Date

observation_period_end_date

Date

period_type_concept_id

integer

...

A data frame with 33 rows and 7 variables:

payer_plan_period_id

integer

person_id

integer

payer_plan_period_start_date

Date

payer_plan_period_end_date

Date

payer_source_value

character

plan_source_value

character

family_source_value

character

...

A data frame with 10 rows and 18 variables:

person_id

integer

gender_concept_id

integer

year_of_birth

integer

month_of_birth

integer

day_of_birth

integer

birth_datetime

POSIXct

race_concept_id

integer

ethnicity_concept_id

integer

location_id

integer

provider_id

integer

care_site_id

integer

person_source_value

character

gender_source_value

character

gender_source_concept_id

integer

race_source_value

character

race_source_concept_id

integer

ethnicity_source_value

character

ethnicity_source_concept_id

integer

...

A data frame with 987 rows and 13 variables:

procedure_occurrence_id

integer

person_id

integer

procedure_concept_id

integer

procedure_date

Date

procedure_datetime

POSIXct

procedure_type_concept_id

integer

modifier_concept_id

integer

quantity

integer

provider_id

integer

visit_occurrence_id

integer

procedure_source_value

character

procedure_source_concept_id

integer

qualifier_source_value

character

...

A data frame with 0 rows and 13 variables:

provider_id

character

provider_name

character

NPI

character

DEA

character

specialty_concept_id

character

care_site_id

character

year_of_birth

character

gender_concept_id

character

provider_source_value

character

specialty_source_value

character

specialty_source_concept_id

character

gender_source_value

character

gender_source_concept_id

character

...

A data frame with 0 rows and 6 variables:

relationship_id

character

relationship_name

character

is_hierarchical

character

defines_ancestry

character

reverse_relationship_id

character

relationship_concept_id

character

...

A data frame with 409 rows and 17 variables:

visit_occurrence_id

integer

person_id

integer

visit_concept_id

integer

visit_start_date

Date

visit_start_datetime

POSIXct

visit_end_date

Date

visit_end_datetime

POSIXct

visit_type_concept_id

integer

provider_id

integer

care_site_id

integer

visit_source_value

character

visit_source_concept_id

integer

admitting_source_concept_id

integer

admitting_source_value

character

discharge_to_concept_id

integer

discharge_to_source_value

character

preceding_visit_occurrence_id

integer

...

A data frame with 0 rows and 5 variables:

vocabulary_id

character

vocabulary_name

character

vocabulary_reference

character

vocabulary_version

character

vocabulary_concept_id

character

...

Source

https://ohdsi.github.io/CommonDataModel/cdm531.html#care_site

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#concept

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#concept_class

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#concept_synonym

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#condition_era

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#condition_occurrence

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#death

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#device_exposure

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#domain

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#drug_era

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#drug_exposure

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#drug_strength

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#measurement

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#note

https://www.mtsamples.com/

https://ohdsi.github.io/CommonDataModel/cdm531.html#observation

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#observation_period

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#payer_plan_period

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#person

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#procedure_occurrence

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#provider

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#relationship

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#visit_occurrence

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF

https://ohdsi.github.io/CommonDataModel/cdm531.html#vocabulary

https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF