MIMIC-III Tables

A collection of functions to create prearranged views of MIMIC-III patient data when supplied with database connection information and a mapping of the connected database.

Usage

mimic3_table_all_patients(table_map, db_connection)

mimic3_table_admissions(table_map, db_connection, subject_id)

mimic3_table_callout(table_map, db_connection, subject_id)

mimic3_table_chart_events(table_map, db_connection, subject_id)

mimic3_table_cpt_events(table_map, db_connection, subject_id)

mimic3_table_diagnoses_icd(table_map, db_connection, subject_id)

mimic3_table_drg_codes(table_map, db_connection, subject_id)

mimic3_table_icu_stays(table_map, db_connection, subject_id)

mimic3_table_lab_events(table_map, db_connection, subject_id)

mimic3_table_microbiology_events(table_map, db_connection, subject_id)

mimic3_table_note_events(table_map, db_connection, subject_id)

mimic3_table_prescriptions(table_map, db_connection, subject_id)

mimic3_table_procedure_events(table_map, db_connection, subject_id)

mimic3_table_procedures_icd(table_map, db_connection, subject_id)

mimic3_table_services(table_map, db_connection, subject_id)

mimic3_table_transfers(table_map, db_connection, subject_id)

Arguments

table_map

A dplyr::tibble containing a mapping between the CDM standard tables and fields to the user connected tables and fields.

db_connection

A DBI::dbConnect object.

subject_id

A numeric, or coercible to numeric.

Value

A dplyr::tibble containing pre-coordinated patient information from the connected database.