Abstraction Module Selector

This module allows the user to select an available ReviewR abstraction module from a dropdown list. It dynamically returns the abstraction setup and instrument user interfaces as well as collected chart abstraction information from the selected module.

This module consists of the following components:

Module UI functions

abstraction_setup_ui: A tagList with a dropdown selector containing available abstraction modules.

abstraction_instrument_ui: A uiOutput containing the data collection instrument UI from the selected chart abstraction module.

Module Server function

abstraction_setup_server: Processes user selection and dynamically returns uiOutputs for the selected abstraction module's setup and data collection instrument user interfaces. Any returns from the configured abstraction module are captured and returned.

abstraction_setup_ui(id)

abstraction_instrument_ui(id)

abstraction_setup_server(id, subject_id)

Arguments

id
The Module namespace

subject_id
A reactive expression containing a subject identifier

Value

abstraction_setup_ui:

tagList

A tagList containing a selectInput that allows for selection of available abstraction setup modules and the setup UI for the selected abstraction module.

abstraction_instrument_ui:
tagList

A tagList containing the selected abstraction module's data collection instrument UI.

abstraction_setup_server:
reactiveValues

This module has no returns of its own, but will pass on the reactiveValues returns from the user selected abstraction module.