Patient Navigation Module

This module will render the "all patients" dataTable (DT) located on the 'Patient Search' tab of ReviewR and will display demographic information about subjects in the connected database. The subject id of the selections made on this tab are extracted and passed to other ReviewR modules. As selections are made using the DT or the navigation buttons on the 'Chart Review' tab, the selected patient in the DT is kept in sync by this module.

Additionally, demographic and (optionally) abstraction status information about the selected patient are extracted and placed into a header on the 'Chart Review' tab.

This module consists of the following components:

Module UI functions

navigation_message: A uiOutput to display a placeholder message when no database is connected.

all_patient_search_dt: A uiOutput containing the "all patients" dataTable with patient demographic information from the connected database.

chart_review_subject_info: A uiOutput containing the selected subject's demographic information to display on the 'Chart Review' tab.

chart_review_navigation: A uiOutput containing the "Jump to Subject ID' dropdown and previous and next buttons used to navigate through patient data on the 'Chart Review' tab.

Module Server function

navigation_server: Provides all of the logic associated with displaying patient demographic information, selecting a patient to review, and navigating through the connected patient database. Returns user selected patient information for use by other modules.

Keyboard Shortcuts

This module also provides keyboard shortcuts to assist with navigating through patient data. The "meta" key refers to "ctrl" on Windows and "Cmd" on Mac.

Next Patient: “alt + meta + >”

Prev Patient: “alt + meta + <”

navigation_message(id)

all_patient_search_dt(id)

chart_review_subject_info(id)

chart_review_navigation(id)

navigation_server(
  id,
  database_vars,
  data_model_vars,
  abstract_vars,
  parent_session
)

Arguments

id
The Module namespace

database_vars
A reactiveValues object as returned by mod_database_setup.

data_model_vars
A reactiveValues object as returned by mod_data_model_detection.

abstract_vars
A reactiveValues object as returned by mod_abstraction_setup.

parent_session
The session information from the parent environment of this module.

Value

navigation_message:

tagList

A uiOutput to display a placeholder message when no database is connected.

all_patient_search_dt:
tagList

A uiOutput containing the "all patients" dataTable with patient demographic information from the connected database.

chart_review_subject_info:
tagList

A uiOutput containing the selected subject's demographic information to display on the 'Chart Review' tab.

chart_review_navigation:
tagList

A uiOutput containing the "Jump to Subject ID' dropdown and previous and next buttons used to navigate through patient data on the 'Chart Review' tab.

navigation_server:
reactiveValues

selected_subject_id: A character representing the currently selected subject in the currently connected database.

selected_subject_info: A dplyr::tibble containing demographic information about the selected subject.

selected_subject_status: A character containing the abstraction status of the selected subject, when an abstraction module is configured and abstraction data is present.