Patient Navigation

This module will render the datatable on the 'Patient Search' tab containing all patients in the cohort. The selected patient in the DT is kept in sync with the 'Chart Review' tab.

navigation_message(id)

mod_navigation_server(
  id,
  database_vars,
  datamodel_vars,
  abstract_vars,
  parent_session
)

Arguments

id
The namespace id for the UI output

database_vars
Database variables returned from user selected database setup module

datamodel_vars
Datamodel variables returned from mod_datamodel_setup

abstract_vars
Abstraction variables returned from user selected abstraction module

parent_session
the parent environment of this module